+++ src/libtcod/engine/display.h
@@ -170,9 +170,25 @@ TCODLIB_API void init_root(
  *
  *  Returns 0 on success, or -1 on an error, you can check the error with
  *  TCOD_sys_get_error()
+ *
+ *  `renderer` and vsync settings can be overridden by the `TCOD_RENDERER` or
+ *  `TCOD_VSYNC` environment variables.
+ *
+ *  Valid case-sensitive options for `TCOD_RENDERER` are:
+ *  - sdl
+ *  - opengl
+ *  - glsl
+ *  - sdl2
+ *  - opengl2
+ *
+ *  Valid options for `TCOD_VSYNC` are `0` or `1`.
+ *
  *  \rst
  *  .. versionchanged:: 1.12
  *      Now returns -1 on error instead of crashing.
+ *
+ *  .. versionchanged:: 1.13
+ *      Added the `TCOD_RENDERER` and `TCOD_VSYNC` overrides.
  *  \endrst
  */
 TCODLIB_CAPI int TCOD_console_init_root(int w, int h, const char* title,